test: 리팩토링 이후 테스트 코드 수정

+++ src/main/java/com/plzgraduate/myongjigraduatebe/graduation/domain/model/DetailGraduationResult.java
@@ -16,7 +16,7 @@ public class DetailGraduationResult {
 
 	@Builder
 	private DetailGraduationResult(GraduationCategory graduationCategory, boolean isCompleted, int totalCredit,
-		int takenCredit,
+		double takenCredit,
 		List<DetailCategoryResult> detailCategory) {
 		this.graduationCategory = graduationCategory;
 		this.isCompleted = isCompleted;
